backend/pdf-service/requirements_ml.txt
# NEXA ML-Specific Requirements
# Separate from main requirements for Docker layer caching
# Last Updated: October 11, 2025

# PyTorch - Auto-detect CPU/GPU based on environment
# For CPU-only: append +cpu to versions
# For CUDA 11.8: append +cu118 to versions
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
# Install with: pip install torch --index-url https://download.pytorch.org/whl/cpu  # for CPU
# Or: pip install torch --index-url https://download.pytorch.org/whl/cu118  # for CUDA

# Hugging Face Ecosystem
transformers==4.35.0
datasets==2.14.0
tokenizers==0.15.0
sentence-transformers==2.2.2
accelerate==0.24.0
peft==0.6.0

# Computer Vision
ultralytics==8.0.200
opencv-python-headless==4.8.0.74
albumentations==1.3.1
roboflow==1.1.9
Pillow==10.0.0
imageio==2.31.1

# NLP & Text Processing
nltk==3.8.1
spacy==3.6.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl
seqeval==1.2.2

# ML Utilities
scikit-learn==1.3.0
scipy==1.11.1
pandas==2.0.3
numpy==1.24.3

# System Monitoring
psutil==5.9.5  # CPU/Memory monitoring
GPUtil==1.4.0  # GPU monitoring
py3nvml==0.2.7  # NVIDIA management library

# Training & Monitoring
tensorboard==2.15.0
tensorboardX==2.6.2
wandb==0.15.12
tqdm==4.66.1
matplotlib==3.7.2
seaborn==0.12.2

# Data Validation
great-expectations==0.17.12
pandera==0.16.1

# Model Optimization
onnx==1.14.1
onnxruntime==1.16.0
openvino==2023.1.0

# Distributed Training & Optimization
deepspeed==0.15.1  # ZeRO optimization for large models
ninja==1.11.1  # Build acceleration for DeepSpeed
py-cpuinfo==9.0.0  # CPU info for DeepSpeed

# Testing ML Code
pytest-ml==0.2.0
hypothesis==6.87.0
networkx==3.3  # For dependency graph analysis - pinned for reliability
pytest-cov==4.1.0  # Coverage reporting
